requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
discord.py==2.3.0
google.generativeai==0.11.0
discord.ext==2.3.0
python-dotenv==1.0.0